Add global + model based transformers.
Currently the only way to set a transform is on each and every request.
This allows for configuration of transforms at the global and model levels.

This is reusing the transform nomenclature but I believe it is most accurate
to describe a change on the raw request or response data.  I think the next
obvious change would be to use the same paradigm to set transforms and interceptors
on a per-action basis (replacing `transformRequest` and `transformResponse`)
to make the nomenclature less confusing.